+++ src/roman_datamodels/stnode/_converters.py
@@ -20,6 +20,8 @@ class _RomanConverter(Converter):
     Base class for the roman_datamodels converters.
     """
 
+    lazy = True
+
     def __init_subclass__(cls, **kwargs) -> None:
         """
         Automatically create the converter objects.
@@ -50,7 +52,7 @@ class TaggedObjectNodeConverter(_RomanConverter):
         return obj.tag
 
     def to_yaml_tree(self, obj, tag, ctx):
-        return obj._data
+        return dict(obj._data)
 
     def from_yaml_tree(self, node, tag, ctx):
         return OBJECT_NODE_CLASSES_BY_TAG[tag](node)